Implement all code review comments and integer-cent salary system
## Code Review Comments (1-7) - Implemented

1. **Case-insensitive metrics queries**: Fixed GROUP BY clauses
   - findMetricsByCountry: GROUP BY LOWER(country), SELECT MIN(country) AS country
   - findAverageSalaryByJobTitle: GROUP BY LOWER(job_title), SELECT MIN(job_title)

2. **Pagination safety bounds**: Added MAX_PAGE_SIZE limit
   - Clamps limit with Math.min(100, Math.max(1, limit))

3. **Eliminated duplicate error handling**: Centralized 404 responses
   - Added status = 404 to NotFoundError
   - Removed 7 duplicate catch blocks from controller
   - Error handler automatically returns 404

4. **Moved database migrations**: Separation of concerns
   - Removed from app.ts
   - Added to server.ts with try-catch

5. **Graceful shutdown handlers**: Production-ready lifecycle
   - Added SIGTERM/SIGINT listeners
   - Closes server, database, exits cleanly

6. **Environment-specific database paths**: Flexible configuration
   - Development: salary_management_dev.db
   - Test: :memory:
   - Production: configurable via DB_PATH

7. **Security middleware**: Production hardening
   - helmet, cors, express-rate-limit
   - Rate limiting skipped in tests

## Integer-Cent Salary System

### Database: salary REAL → INTEGER (cents)
- Stores values * 100 (755.00 = 75500)
- Eliminates floating-point precision errors

### Input: Validators transform decimal → cents
- salary field: .transform(val => Math.round(val * 100))

### Output: Serialization converts cents → decimal
- serializeEmployee(): Returns salary / 100
- serializeMetrics functions: Converts metrics back to decimal

### Calculations: Internal operations on cents
- calculateDeductions(grossSalaryInCents)
- calculateSalary converts to decimal in response

## Tests: All 41 Passing ✅
- employee.crud.test.ts: PASS
- salary.metrics.test.ts: PASS
- salary.calculator.unit.test.ts: PASS
- salary.calculation.test.ts: PASS

Build: No TypeScript errors ✅

+++ src/database/migrations.ts
@@ -9,7 +9,7 @@ export function runMigrations(): void {
       full_name TEXT NOT NULL,
       job_title TEXT NOT NULL,
       country TEXT NOT NULL,
-      salary REAL NOT NULL CHECK(salary > 0),
+      salary INTEGER NOT NULL CHECK(salary > 0),
       created_at TEXT NOT NULL DEFAULT (datetime('now')),
       updated_at TEXT NOT NULL DEFAULT (datetime('now'))
     );
